+++ tests/components/squeezebox/snapshots/test_media_player.ambr
@@ -1,82 +1,4 @@
 # serializer version: 1
-# name: test_device_registry
-  DeviceRegistryEntrySnapshot({
-    'area_id': None,
-    'config_entries': <ANY>,
-    'config_entries_subentries': <ANY>,
-    'configuration_url': None,
-    'connections': set({
-      tuple(
-        'mac',
-        'aa:bb:cc:dd:ee:ff',
-      ),
-    }),
-    'disabled_by': None,
-    'entry_type': None,
-    'hw_version': None,
-    'id': <ANY>,
-    'identifiers': set({
-      tuple(
-        'squeezebox',
-        'aa:bb:cc:dd:ee:ff',
-      ),
-    }),
-    'is_new': False,
-    'labels': set({
-    }),
-    'manufacturer': 'Ralph Irving & Adrian Smith',
-    'model': 'SqueezeLite',
-    'model_id': None,
-    'name': 'Test Player',
-    'name_by_user': None,
-    'primary_config_entry': <ANY>,
-    'serial_number': None,
-    'suggested_area': None,
-    'sw_version': '',
-    'via_device_id': <ANY>,
-  })
-# ---
-# name: test_device_registry_server_merged
-  DeviceRegistryEntrySnapshot({
-    'area_id': None,
-    'config_entries': <ANY>,
-    'config_entries_subentries': <ANY>,
-    'configuration_url': None,
-    'connections': set({
-      tuple(
-        'mac',
-        'ff:ee:dd:cc:bb:aa',
-      ),
-    }),
-    'disabled_by': None,
-    'entry_type': <DeviceEntryType.SERVICE: 'service'>,
-    'hw_version': None,
-    'id': <ANY>,
-    'identifiers': set({
-      tuple(
-        'squeezebox',
-        '12345678-1234-1234-1234-123456789012',
-      ),
-      tuple(
-        'squeezebox',
-        'ff:ee:dd:cc:bb:aa',
-      ),
-    }),
-    'is_new': False,
-    'labels': set({
-    }),
-    'manufacturer': 'https://lyrion.org/ / Ralph Irving & Adrian Smith',
-    'model': 'Lyrion Music Server/SqueezeLite',
-    'model_id': 'LMS',
-    'name': '1.2.3.4',
-    'name_by_user': None,
-    'primary_config_entry': <ANY>,
-    'serial_number': None,
-    'suggested_area': None,
-    'sw_version': '',
-    'via_device_id': <ANY>,
-  })
-# ---
 # name: test_entity_registry[media_player.test_player-entry]
   EntityRegistryEntrySnapshot({
     'aliases': set({
