+++ tests/components/squeezebox/test_init.py
@@ -1,10 +1,16 @@
 """Test squeezebox initialization."""
 
 from http import HTTPStatus
-from unittest.mock import patch
+from unittest.mock import MagicMock, patch
 
+from syrupy.assertion import SnapshotAssertion
+
+from homeassistant.components.squeezebox.const import DOMAIN
 from homeassistant.config_entries import ConfigEntryState
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.device_registry import DeviceRegistry
+
+from .conftest import TEST_MAC
 
 from tests.common import MockConfigEntry
 
@@ -82,3 +88,27 @@ async def test_init_missing_uuid(
         mock_async_query.assert_called_once_with(
             "serverstatus", "-", "-", "prefs:libraryname"
         )
+
+
+async def test_device_registry(
+    hass: HomeAssistant,
+    device_registry: DeviceRegistry,
+    configured_player: MagicMock,
+    snapshot: SnapshotAssertion,
+) -> None:
+    """Test squeezebox device registered in the device registry."""
+    reg_device = device_registry.async_get_device(identifiers={(DOMAIN, TEST_MAC[0])})
+    assert reg_device is not None
+    assert reg_device == snapshot
+
+
+async def test_device_registry_server_merged(
+    hass: HomeAssistant,
+    device_registry: DeviceRegistry,
+    configured_players: MagicMock,
+    snapshot: SnapshotAssertion,
+) -> None:
+    """Test squeezebox device registered in the device registry."""
+    reg_device = device_registry.async_get_device(identifiers={(DOMAIN, TEST_MAC[2])})
+    assert reg_device is not None
+    assert reg_device == snapshot
